refactor: route admin storage facades
