+++ rustfs/src/admin/handlers/quota.rs
@@ -14,12 +14,12 @@
 
 //! Quota admin handlers for HTTP API
 
-use super::super::metadata_sys::BucketMetadataSys;
-use super::super::quota::checker::QuotaChecker;
-use super::super::quota::{BucketQuota, QuotaError, QuotaOperation};
 use crate::admin::auth::{validate_admin_request, validate_admin_request_with_bucket};
 use crate::admin::router::{AdminOperation, Operation, S3Router};
 use crate::admin::runtime_sources::{resolve_bucket_metadata_handle, resolve_object_store_handle};
+use crate::admin::storage_api::metadata_sys::BucketMetadataSys;
+use crate::admin::storage_api::quota::checker::QuotaChecker;
+use crate::admin::storage_api::quota::{BucketQuota, QuotaError, QuotaOperation};
 use crate::auth::{check_key_valid, get_session_token};
 use crate::server::ADMIN_PREFIX;
 use hyper::{Method, StatusCode};
@@ -175,7 +175,7 @@ async fn current_usage_from_context(bucket: &str) -> u64 {
         return 0;
     };
 
-    match super::super::storage_api::data_usage::load_data_usage_from_backend(store).await {
+    match crate::admin::storage_api::data_usage::load_data_usage_from_backend(store).await {
         Ok(data_usage_info) => data_usage_info
             .buckets_usage
             .get(bucket)
