+++ rustfs/src/admin/handlers/tier.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 #![allow(unused_variables, unused_mut, unused_must_use)]
 
-use super::super::{
+use crate::admin::storage_api::{
     AdminError, DailyAllTierStats, ERR_TIER_ALREADY_EXISTS, ERR_TIER_BACKEND_IN_USE, ERR_TIER_BACKEND_NOT_EMPTY,
     ERR_TIER_CONNECT_ERR, ERR_TIER_INVALID_CREDENTIALS, ERR_TIER_MISSING_CREDENTIALS, ERR_TIER_NAME_NOT_UPPERCASE,
     ERR_TIER_NOT_FOUND, TierConfig, TierCreds, TierType, storageclass,
@@ -931,8 +931,8 @@ impl Operation for ClearTier {
 
 #[cfg(test)]
 mod tests {
-    use super::super::super::lifecycle::tier_last_day_stats::LastDayTierStats;
     use super::*;
+    use crate::admin::storage_api::lifecycle::tier_last_day_stats::LastDayTierStats;
     use http::Uri;
     use matchit::Router;
