+++ rustfs/src/admin/handlers/trace.rs
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::PeerRestClient;
 use crate::admin::router::Operation;
 use crate::admin::runtime_sources::resolve_endpoints_handle;
+use crate::admin::storage_api::PeerRestClient;
 use http::StatusCode;
 use hyper::Uri;
 use matchit::Params;
