+++ rustfs/src/admin/service/site_replication.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::Error as StorageError;
-use super::super::{read_admin_config, save_admin_config};
 use crate::admin::runtime_sources::{AppContext, get_global_app_context, resolve_object_store_handle_for_context};
 use crate::admin::site_replication_identity::{deployment_id_for_endpoint, normalize_peer_map_by_identity_with};
+use crate::admin::storage_api::Error as StorageError;
+use crate::admin::storage_api::{read_admin_config, save_admin_config};
 use rustfs_madmin::PeerInfo;
 use s3s::{S3Error, S3ErrorCode, S3Result};
 use serde_json::{Map, Value};
